refactoring/following conventions/fixing packages/fixing architecture
